✨feat: 테스트 케이스 추가 & 패키지 생성
