+++ src/test/java/LadderGameTest/LadderTest.java
@@ -1,0 +1,38 @@
+package LadderGameTest;
+
+import LadderGame.domain.Ladder;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+public class LadderTest {
+    @Test
+    @DisplayName("00. 기본 테스트")
+    void test00() {
+        Ladder ladder = new Ladder("pobi,honux,crong,jk", 5);
+        ladder.makeLadder();
+        ladder.callOutput();
+    }
+
+    @Test
+    @DisplayName("01. 사람 수 테스트(100명)")
+    void test01() {
+        String str = "";
+
+        for(int i=1; i<100; i++){
+            str += Integer.toString(i) + ",";
+        }
+        str += "100";
+
+        Ladder ladder = new Ladder(str, 5);
+        ladder.makeLadder();
+        ladder.callOutput();
+    }
+
+    @Test
+    @DisplayName("02. 사다리 높이 테스트(100)")
+    void test02() {
+        Ladder ladder = new Ladder("pobi,honux,crong,jk", 100);
+        ladder.makeLadder();
+        ladder.callOutput();
+    }
+}
